Moved methods public User getCurrentUser() public Long getCurrentUserId() from CustomUserDetailsService to UserService
