+++ src/main/java/com/example/bookstoreapp/security/CustomUserDetailsService.java
@@ -1,9 +1,7 @@
 package com.example.bookstoreapp.security;
 
-import com.example.bookstoreapp.model.User;
 import com.example.bookstoreapp.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -19,18 +17,5 @@ public class CustomUserDetailsService implements UserDetailsService {
         return userRepository.findByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Can't find user by "
                 + username + " username."));
-    }
-
-    public User getCurrentUser() {
-        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        if (principal instanceof User) {
-            return (User) principal;
-        } else {
-            throw new RuntimeException("User not authenticated");
-        }
-    }
-
-    public Long getCurrentUserId() {
-        return getCurrentUser().getId();
     }
 }
