+++ src/main/java/com/example/bookstoreapp/servive/UserService.java
@@ -3,8 +3,13 @@ package com.example.bookstoreapp.servive;
 import com.example.bookstoreapp.dto.UserRegistrationRequestDto;
 import com.example.bookstoreapp.dto.UserResponseDto;
 import com.example.bookstoreapp.exception.RegistrationException;
+import com.example.bookstoreapp.model.User;
 
 public interface UserService {
     UserResponseDto register(UserRegistrationRequestDto userRegistrationRequestDto)
             throws RegistrationException;
+
+    Long getCurrentUserId();
+
+    User getCurrentUser();
 }
